+++ src/CHANGELOG.md
@@ -15,3 +15,11 @@ All notable changes to this crate will be documented in this file from v0.5.0
 ### Fixed
 - Reconnection errors for the client.
 - Stack overflow bug on layers caused by recursive functions
+
+
+## [0.5.1] - 2026.01.12
+## Added
+- Intercepting (instable)
+  - Interceptor Struct
+  - InterceptorResult Enum (Pass(String) or Cancel)
+  - Intercept func for client and server